requirements.txt
google-generativeai
gradio
opencv-python
gTTS
google-cloud-vision
torch
transformers
moviepy